+++ clUtil/clUtil.cpp
@@ -53,7 +53,7 @@ std::vector<cl::CLDeviceInfo> cl::getDevices()
             clGetDeviceInfo(devices[j], CL_DEVICE_GLOBAL_MEM_SIZE, sizeof(mem), &mem, NULL);
 
             info.mem = (uint64_t)mem;
-            info.id = (uint64_t)devices[j];
+            info.id = devices[j];
             deviceList.push_back(info);
         }
